Hands_on_Lab/Day_3/Task_2/requirements.txt
flask
langgraph
langchain
langchain-groq
langchain-community
transformers
torch
python-dotenv
